refactor(tests): consolidate orchestra test files from 14 to 11
Changes:
- Delete test_global_dispatch_coordinator.py (pure re-export file)
- Merge test_dispatch_queue_operations.py + test_dispatch_actionable_trigger.py
  → test_dispatch_queue_management.py (25 tests)
- Merge test_failed_gate.py + test_failed_gate_integration.py
  → test_failed_gate.py (13 tests)

Result:
- Test files: 14 → 11 (reduced by 3)
- All 126 tests passing
- No lint errors

Rationale:
- test_global_dispatch_coordinator.py only re-exported tests already in other files
- dispatch queue tests were testing the same component with overlapping patterns
- FailedGate unit + integration tests are clearer when in one file
